voting: extract status screen into shared flow and steps body
VotingStatusFlow owns job start scheduling, display derivation, Keystone
scan/skip handoffs, retry/clear, and confirmation navigation;
VotingStatusStepsBody carries the step checklist, Keystone signing panel,
and wallet-sync card. The desktop screen is now a thin shell providing
its scroll container, header copy, and skip-confirmation dialog. No
behavior change.
